backends/default_django/pyproject.toml
[project]
name = "ufrecs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beartype>=0.21.0",
    "boto3>=1.40.16",
    "django~=5.2",
    "django-crispy-forms>=2.4",
    "django-filter>=25.1",
    "django-magic-notifier>=0.4.0",
    "django-silk>=5.4.1",
    "django-storages>=1.14.6",
    "django-stubs-ext>=5.2.2",
    "django-uni-ps",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "pillow>=11.3.0",
    "python-decouple>=3.8",
    "traceback-with-variables>=2.2.0",
]

[tool.uv.sources]
django-uni-ps = { git = "https://gitlab.ngoul.com/yaknema/uni-ps.git" }
